feat: large refactor of the endpoint generation loop to use factories. Fixes the fundamental security bug
