--- src/fides/api/ctl/routes/generic_crud_utils.py
@@ -1,1 +1,0 @@
-# Put the old functions here
