+++ src/fides/api/main.py
@@ -22,7 +22,7 @@ import fides
 from fides.api.ctl import view
 from fides.api.ctl.database.database import configure_db
 from fides.api.ctl.database.seed import create_or_update_parent_user
-from fides.api.ctl.routes import admin, crud, generate, health, system, validate
+from fides.api.ctl.routes import admin, generate, generic, health, system, validate
 from fides.api.ctl.routes.util import API_PREFIX
 from fides.api.ctl.ui import (
     get_admin_index_as_response,
@@ -71,7 +71,7 @@ from fides.core.config import CONFIG, check_required_webserver_config_values
 
 VERSION = fides.__version__
 
-ROUTERS = crud.routers + [  # type: ignore[attr-defined]
+ROUTERS = generic.routers + [  # type: ignore[attr-defined]
     admin.router,
     generate.router,
     health.router,
